Rename processors based on review feedback
