+++ nifi-extension-bundles/nifi-azure-bundle/nifi-azure-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -27,8 +27,8 @@ org.apache.nifi.processors.azure.storage.FetchAzureBlobStorage_v12
 org.apache.nifi.processors.azure.storage.PutAzureBlobStorage_v12
 org.apache.nifi.processors.azure.storage.DeleteAzureBlobStorage_v12
 org.apache.nifi.processors.azure.storage.CopyAzureBlobStorage_v12
-org.apache.nifi.processors.azure.storage.GetAzureBlobStorageMetadata_v12
-org.apache.nifi.processors.azure.storage.GetAzureBlobStorageTags_v12
+org.apache.nifi.processors.azure.storage.GetAzureBlobStorageMetadata
+org.apache.nifi.processors.azure.storage.GetAzureBlobStorageTags
 org.apache.nifi.processors.azure.storage.MoveAzureDataLakeStorage
 org.apache.nifi.processors.azure.storage.queue.GetAzureQueueStorage_v12
 org.apache.nifi.processors.azure.storage.queue.PutAzureQueueStorage_v12
